split off `statement` and `segment` modules (#69)
